Implement features

+++ app/Http/Controllers/v1/ProjectController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers\v1;
 
 use App\Models\Project;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class ProjectController extends Controller
 {
@@ -12,23 +15,18 @@ class ProjectController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index() //GET
+    public function index(Request $request) //GET
     {
-        $projects = Project::all();
-        return [
+        //Team member only have visibility to their projects.
+        if($request->user()->tokenCan('role:team_member')){
+            $projects = Project::whereRaw("json_contains(members,'".$request->user()->id."')")->get();
+        }else{
+            $projects = Project::all();
+        }
+        return response()->json([
             "success" => true,
             "data" => $projects
-        ];
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create() //NOT USED
-    {
-        //
+        ], 200);
     }
 
     /**
@@ -39,20 +37,38 @@ class ProjectController extends Controller
      */
     public function store(Request $request) //POST 
     {
-        $request->validate([
+        if(!$request->user()->tokenCan('role:product_owner')){
+            throw new AccessDeniedHttpException($request);
+        }
+
+        $validate = Validator::make($request->all(), 
+        [
             'name' => 'required',
         ]);
+
+        if($validate->fails()){
+            return response()->json([
+                'success' => false,
+                'msg' => 'Validation error: '.$validate->errors(),
+            ], 401);
+        }
+
         if(Project::where('name',$request->post('name'))->count() > 0){
-            return [
+            return response()->json([
                 'success'=>false,
                 'msg'=>'Project name already taken'
-            ];
+            ], 401);
         }
-        $projects = Project::create($request->all());
-        return [
+        
+        $projects = Project::create([
+            'name'=>$request->name,
+            'members'=>!empty($request->members)?json_encode($request->members):null
+        ]);
+
+        return response()->json([
             "success" => true,
             "data" => $projects
-        ];  
+        ], 200);
     }
 
     /**
@@ -61,23 +77,29 @@ class ProjectController extends Controller
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function show(Project $project) //GET BY ID
+    public function show(Request $request, $id) //GET BY ID
     {
-        return [
+        $project = Project::find($id);
+        if(empty($project)){
+            $project = [];
+        }
+        // Team member only have visibility to their projects.
+        if($request->user()->tokenCan('role:team_member')){
+            $members = json_decode($project->members,true);
+            if(!in_array($request->user()->id,$members)){
+                return response()->json([
+                    "success" => false,
+                    "msg" => "User does not have permission to view this project"
+                ], 401);
+            }
+        }else{
+            $projects = Project::all();
+        }
+        //Team member only have visibility to their projects.
+        return response()->json([
             "success" => true,
             "data" =>$project
-        ];
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Project  $project
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Project $proejct) //NOT IMPLEMENTED
-    {
-        //
+        ], 200);
     }
 
     /**
@@ -87,19 +109,41 @@ class ProjectController extends Controller
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Project $project) //PUT OR PATCH
+    public function update(Request $request, $id) //PUT OR PATCH
     {
-        $request->validate([
+        //Only PRODUCT_OWNER role can create a project and tasks
+        if(!$request->user()->tokenCan('role:product_owner')){
+            throw new AccessDeniedHttpException($request);
+        }
+        $validate = Validator::make($request->all(), 
+        [
             'name' => 'required',
         ]);
+
+        if($validate->fails()){
+            return response()->json([
+                'success' => false,
+                'msg' => 'Validation error: '.$validate->errors(),
+            ], 401);
+        }
+        
+        $project = Project::find($id);
+        if(empty($project)){
+            return response()->json([
+                "success" => false,
+                "msg" => "Project not found"
+            ], 401);
+        }
+        $project->update([
+            'name'=>$request->name,
+            'members'=>!empty($request->members)?$request->members:null
+        ]);
  
-        $project->update($request->all());
- 
-        return [
+        return response()->json([
             "success" => true,
             "data" => $project,
             "msg" => "Project updated successfully"
-        ];
+        ], 200);
     }
 
     /**
@@ -108,19 +152,22 @@ class ProjectController extends Controller
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id) //DELETE
+    public function destroy(Request $request, $id) //DELETE
     {
+        if(!$request->user()->tokenCan('role:product_owner')){
+            throw new AccessDeniedHttpException($request);
+        }
         $ok = Project::where('id', $id)->delete();
         if($ok){
-            return [
+            return response()->json([
                 "success" => true,
                 "msg" => "Project deleted successfully"
-            ];
+            ], 200);
         }else{
-            return [
+            return response()->json([
                 'success' => false,
                 'msg'=>"Unable to delete"
-            ];
+            ], 401);
         }
     }
 }
